Triplanar mapping, slope texturing, new perlin
This commit adds textures for the terrain, right now they are very
strectched in one direction which is a bug for later. There is also a
simple implementation of applying textures depending on the slope
(should be better implemented later).
Re-did the perlin noise generation to be more like the original,
improved perlin noise algorithm